requirements.txt
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
deepl==1.5.0
facebook-sdk==3.1.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==9.0.1
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
urllib3==1.26.8
